+++ src/AccountManagement/Ui/User/Web/Controller/VerifyUserEmailController.php
@@ -1,0 +1,58 @@
+<?php
+
+namespace App\AccountManagement\Ui\User\Web\Controller;
+
+use App\AccountManagement\Domain\User\Model\User;
+use App\Security\EmailVerifier;
+use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Contracts\Translation\TranslatorInterface;
+use SymfonyCasts\Bundle\VerifyEmail\Exception\VerifyEmailExceptionInterface;
+
+class VerifyUserEmailController extends AbstractController
+{
+    private EmailVerifier $emailVerifier;
+
+    public function __construct(EmailVerifier $emailVerifier)
+    {
+        $this->emailVerifier = $emailVerifier;
+    }
+
+    #[Route('/verify/email', name: 'app_verify_email')]
+    public function __invoke(
+        Request                $request,
+        TranslatorInterface    $translator,
+        EntityManagerInterface $manager
+    ): Response
+    {
+        $id = $request->get('id');
+
+        if (null === $id) {
+            return $this->redirectToRoute('register');
+        }
+
+        $repository = $manager->getRepository(User::class);
+        $user = $repository->find($id);
+
+        if (null === $user) {
+            return $this->redirectToRoute('register');
+        }
+
+        // validate email confirmation link, sets User::isVerified=true and persists
+        try {
+            $this->emailVerifier->handleEmailConfirmation($request, $user);
+        } catch (VerifyEmailExceptionInterface $exception) {
+            $this->addFlash('verify_email_error', $translator->trans($exception->getReason(), [], 'VerifyEmailBundle'));
+
+            return $this->redirectToRoute('register');
+        }
+
+        // @TODO Change the redirect on success and handle or remove the flash message in your templates
+        $this->addFlash('success', 'Your email address has been verified.');
+
+        return $this->redirectToRoute('register');
+    }
+}
